requirements.txt
# Add your Python dependencies here, one per line.

pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2 # for machine learning
matplotlib==3.8.0
seaborn==0.12.2
scipy==1.11.4  # for scientific computing
requests==2.31.0
python-dotenv==1.0.0  # for environment variable management
reportlab==3.6.12  # for PDF generation
Pillow  # optional, for images
beautifulsoup4==4.12.2  # for web scraping 
openai==1.0.0  # for OpenAI API interactions 
selenium==4.20.0  # for web automation 




